Add files via upload
A simple game of two player battleship. The spot is randomly assigned and it is up to two players to guess where they should sink their battleship. It will keep going until the players do not want to play.

This was built upon the code given to me by codeacademy.